refactor: diviser les mails dans les fichiers séparés
